requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.0
openai==1.3.0
anthropic==0.5.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
prometheus-client>=0.19.0
redis>=5.0.0
httpx>=0.24.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0
slowapi>=0.1.9